+++ python_blockchain_app-ibm_blockchain_post-2/app/auth.py
@@ -53,4 +53,4 @@ def signup_post():
 @login_required
 def logout():
     logout_user()
-    return redirect(url_for('main.index'))
+    return redirect(url_for('main.sample'))
