--- python_blockchain_app-ibm_blockchain_post-2/app/templates/base(original).html
@@ -1,20 +1,0 @@
-<html>
-  <head>
-    <title>Add Review</title>
-  </head>
-  <body>
-    <div><a href="/">Return to home</a></div>
-    <center><h1>Submit Anonymous Review</h1></center>
-    <hr>
-    {% with messages = get_flashed_messages() %}
-    {% if messages %}
-    <ul>
-    {% for message in messages %}
-        <li>{{ message }} </li>
-    {% endfor %}
-    </ul>
-    {% endif %}
-    {% endwith %}
-    {% block content %}{% endblock %}
-  </body>
-</html>
