+++ python_blockchain_app-ibm_blockchain_post-2/app/templates/login.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "base2.html" %}
 
 {% block content %}
 <div class="column is-4 is-offset-4">
