+++ python_blockchain_app-ibm_blockchain_post-2/app/templates/profile.html
@@ -1,4 +1,4 @@
-{% extends "base.html" %}
+{% extends "base2.html" %}
 
 {% block content %}
 <h1 class="title">
